+++ webdev/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.2.0-dev
+
+- Added a `fullReload` service extension which performs a full page refresh.
+
 ## 2.1.0
 
 - Add an explicit error if there are no directories to serve. Typically this